Split large 'deploy' integration test
* remove a file called util_*, move test data to the catalog pkg
* added machine prefix to instance group related test clients file name,
  to match existing file names
* moved container env key helper into machine pkg, where some of those
  helpers already reside
