+++ integration/util/tlogs_test.go
@@ -25,7 +25,7 @@ var _ = Describe("when testing tail-logs subcommand", func() {
 			do echo "nats-msg-line" >> /var/vcap/sys/log/nats/nats.log; sleep 5;
 			done`)
 
-			testPod := env.CreatePodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
+			testPod := env.PodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
 
 			tearDown, err := env.CreatePod(env.Namespace, testPod)
 			Expect(err).NotTo(HaveOccurred())
@@ -51,7 +51,7 @@ var _ = Describe("when testing tail-logs subcommand", func() {
 			echo "doppler-msg-line" >> /var/vcap/sys/log/doppler/doppler.log; sleep 5;
 			done`)
 
-			testPod := env.CreatePodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
+			testPod := env.PodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
 
 			tearDown, err := env.CreatePod(env.Namespace, testPod)
 			Expect(err).NotTo(HaveOccurred())
@@ -76,7 +76,7 @@ var _ = Describe("when testing tail-logs subcommand", func() {
 			echo "nats-error-msg-line" >> /var/vcap/sys/log/nats/nats.err; sleep 5;
 			done`)
 
-			testPod := env.CreatePodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
+			testPod := env.PodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
 
 			tearDown, err := env.CreatePod(env.Namespace, testPod)
 			Expect(err).NotTo(HaveOccurred())
@@ -96,7 +96,7 @@ var _ = Describe("when testing tail-logs subcommand", func() {
 			do sleep 5;
 			done`)
 
-			testPod := env.CreatePodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
+			testPod := env.PodWithTailLogsContainer(podName, scriptCreateDirs, parentCName, sidecarCName, operatorimage.GetOperatorDockerImage())
 
 			tearDown, err := env.CreatePod(env.Namespace, testPod)
 			Expect(err).NotTo(HaveOccurred())
