+++ testing/catalog.go
@@ -16,6 +16,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
 
+	"code.cloudfoundry.org/cf-operator/pkg/bosh/bpmconverter"
 	"code.cloudfoundry.org/cf-operator/pkg/bosh/manifest"
 	"code.cloudfoundry.org/cf-operator/pkg/credsgen"
 	bdv1 "code.cloudfoundry.org/cf-operator/pkg/kube/apis/boshdeployment/v1alpha1"
@@ -30,6 +31,8 @@ import (
 
 const (
 	manifestFailedMessage = "Loading bosh manifest spec failed."
+	logsDir               = "LOGS_DIR"
+	logsDirPath           = "/var/vcap/sys/log"
 )
 
 // Catalog provides several instances for tests
@@ -808,6 +811,68 @@ func (c *Catalog) Sleep1hPodSpec() corev1.PodSpec {
 				Name:    "busybox",
 				Image:   "busybox",
 				Command: []string{"sleep", "3600"},
+			},
+		},
+	}
+}
+
+// PodWithTailLogsContainer will generate a pod with two containers
+// One is the parent container that will execute a cmd, preferrable something
+// that writes into files under /var/vcap/sys/log
+// The side-car container, will be tailing the logs of specific files under
+// /var/vcap/sys/log, by running the cf-operator util tail-logs subcmommand
+func (c *Catalog) PodWithTailLogsContainer(podName string, parentPodCmd string, parentCName string, sidecardCName string,
+	dockerImg string) corev1.Pod {
+	return corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: podName,
+		},
+		Spec: corev1.PodSpec{
+			RestartPolicy: corev1.RestartPolicyOnFailure,
+			Containers: []corev1.Container{
+				{
+					Name:  parentCName,
+					Image: dockerImg,
+					VolumeMounts: []corev1.VolumeMount{
+						{
+							Name:      bpmconverter.VolumeSysDirName,
+							MountPath: bpmconverter.VolumeSysDirMountPath,
+						},
+					},
+					Command: []string{
+						"/bin/sh",
+					},
+					Args: []string{
+						"-xc",
+						parentPodCmd,
+					},
+				},
+				{
+					Name:  sidecardCName,
+					Image: dockerImg,
+					VolumeMounts: []corev1.VolumeMount{
+						{
+							Name:      bpmconverter.VolumeSysDirName,
+							MountPath: bpmconverter.VolumeSysDirMountPath,
+						},
+					},
+					Args: []string{
+						"util",
+						"tail-logs",
+					},
+					Env: []corev1.EnvVar{
+						{
+							Name:  logsDir,
+							Value: logsDirPath,
+						},
+					},
+				},
+			},
+			Volumes: []corev1.Volume{
+				{
+					Name:         bpmconverter.VolumeSysDirName,
+					VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}},
+				},
 			},
 		},
 	}
